Time: 11 ms (46.54%) | Memory: 22.8 MB (58.00%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -5,7 +5,7 @@ public:
         while(l<r)
         {
             mid = l+(r-l)/2;
-            long long x = 0;
+            int x = 0;
             for(auto i :piles)
             {
                 x+=(i+mid-1)/mid;
